+++ spynnaker_extra_pynn_models/neural_models/abstract_delta_population_vertex.py
@@ -1,124 +1,48 @@
-"""
-IFCurrentDeltaPopulation
-"""
-from spynnaker.pyNN.models.abstract_models.abstract_population_vertex import \
-    AbstractPopulationVertex
-from spynnaker.pyNN.utilities import constants
-from abstract_delta_population_vertex import AbstractDeltaPopulationVertex
-from spynnaker.pyNN.models.abstract_models.abstract_model_components.\
-    abstract_integrate_and_fire_properties \
-    import AbstractIntegrateAndFireProperties
-from spynnaker.pyNN.models.neural_properties.neural_parameter \
-    import NeuronParameter
+from spynnaker.pyNN.utilities.constants import POPULATION_BASED_REGIONS
+from spynnaker.pyNN.utilities import utility_calls
+from spynnaker.pyNN.models.utility_models.exp_synapse_param\
+    import write_exp_synapse_param
+from abc import ABCMeta
+from six import add_metaclass
+from abc import abstractmethod
 
 
-from data_specification.enums.data_type import DataType
-
-
-class IFCurrentDeltaPopulation(AbstractDeltaPopulationVertex,
-                                     AbstractIntegrateAndFireProperties,
-                                     AbstractPopulationVertex):
+@add_metaclass(ABCMeta)
+class AbstractDeltaPopulationVertex(object):
     """
-    IFCurrentExponentialPopulation: model which represents a leaky intergate
-    and fire model with a exponetial decay curve and based off current.
+    This represents a pynn_population.py with two delta synapses
     """
-
-    _model_based_max_atoms_per_core = 256
-
-    # noinspection PyPep8Naming
-    def __init__(self, n_neurons, machine_time_step, timescale_factor,
-                 spikes_per_second, ring_buffer_sigma, constraints=None,
-                 label=None, tau_m=20.0, cm=1.0, v_rest=-65.0, v_reset=-65.0,
-                 v_thresh=-50.0, tau_refrac=0.1,
-                 i_offset=0, v_init=None):
-        # Instantiate the parent classes
-        AbstractIntegrateAndFireProperties.__init__(
-            self, atoms=n_neurons, cm=cm, tau_m=tau_m, i_offset=i_offset,
-            v_init=v_init, v_reset=v_reset, v_rest=v_rest, v_thresh=v_thresh,
-            tau_refrac=tau_refrac)
-        AbstractPopulationVertex.__init__(
-            self, n_neurons=n_neurons, n_params=10, label=label,
-            binary="IF_curr_delta.aplx", constraints=constraints,
-            max_atoms_per_core=(IFCurrentExponentialPopulation
-                                ._model_based_max_atoms_per_core),
-            machine_time_step=machine_time_step,
-            timescale_factor=timescale_factor,
-            spikes_per_second=spikes_per_second,
-            ring_buffer_sigma=ring_buffer_sigma)
-
-    @property
-    def model_name(self):
-        """
-
+    @abstractmethod
+    def is_delta_vertex(self):
+        """helper method for is_instance
         :return:
         """
-        return "IF_curr_delta"
+
+    def get_n_synapse_parameters_per_synapse_type(self):
+
+        # There are 2 synapse parameters per synapse type (tau_syn and initial)
+        return 0
+
+    def get_n_synapse_types(self):
+
+        # There are 2 synapse types (excitatory and inhibitory)
+        return 2
 
     @staticmethod
-    def set_model_max_atoms_per_core(new_value):
+    def get_n_synapse_type_bits():
+        """
+        Return the number of bits used to identify the synapse in the synaptic
+        row
+        """
+        return 1
+
+    def write_synapse_parameters(self, spec, subvertex, vertex_slice):
+        """
+        **YUCK** Basically does nothing beside switch region
         """
 
-        :param new_value:
-        :return:
-        """
-        IFCurrentDeltaPopulation.\
-            _model_based_max_atoms_per_core = new_value
+        # Set the focus to the memory region 3 (synapse parameters):
+        spec.switch_write_focus(
+            region=POPULATION_BASED_REGIONS.SYNAPSE_PARAMS.value)
 
-    def get_cpu_usage_for_atoms(self, vertex_slice, graph):
-        """
-
-        :param vertex_slice:
-        :param graph:
-        :return:
-        """
-        return 781 * ((vertex_slice.hi_atom - vertex_slice.lo_atom) + 1)
-
-    def get_parameters(self):
-        """
-        Generate Neuron Parameter data (region 2):
-        """
-        # Get the parameters
-        return [
-            NeuronParameter(self._v_thresh, DataType.S1615),
-            NeuronParameter(self._v_reset, DataType.S1615),
-            NeuronParameter(self._v_rest, DataType.S1615),
-            NeuronParameter(self.r_membrane(self._machine_time_step),
-                            DataType.S1615),
-            NeuronParameter(self._v_init, DataType.S1615),
-            NeuronParameter(self.ioffset(self._machine_time_step),
-                            DataType.S1615),
-            NeuronParameter(self.exp_tc(self._machine_time_step),
-                            DataType.S1615),
-            NeuronParameter(self._one_over_tau_rc, DataType.S1615),
-            NeuronParameter(self._refract_timer, DataType.INT32),
-            # t refact used to be a uint32 but was changed to int32 to avoid
-            # clash of c and python variable typing.
-            NeuronParameter(self._scaled_t_refract(), DataType.INT32)]
-
-    def is_population_vertex(self):
-        """
-
-        :return:
-        """
-        return True
-
-    def is_integrate_and_fire_vertex(self):
-        """
-
-        :return:
-        """
-        return True
-
-    def is_delta_vertex(self):
-        """
-
-        :return:
-        """
-        return True
-
-    def is_recordable(self):
-        """
-
-        :return:
-        """
-        return True
+        spec.comment("\nWriting empty delta synapse parameters")
